requirements.txt
# FastAPI & Server
fastapi>=0.115.0
uvicorn[standard]>=0.34.0
python-multipart>=0.0.18
pydantic-settings>=2.7.0

# Database
sqlalchemy[asyncio]>=2.0.36
asyncpg>=0.30.0

# Auth
python-jose[cryptography]>=3.3.0

# AI APIs
google-genai>=1.14.0
openai>=1.58.0
httpx>=0.28.0

# Video/Image Processing
Pillow>=11.0.0
ffmpeg-python>=0.2.0

# YouTube API
google-api-python-client>=2.150.0
google-auth>=2.37.0

# Task Queue & Scheduling
apscheduler>=3.10.4

# Utilities
python-dotenv>=1.0.1
aiofiles>=24.1.0
jinja2>=3.1.4
pydantic-settings==2.7.0
